+++ modules/client/test/src/com/worxbend/codeberg4s/AttemptParitySuite.scala
@@ -122,6 +122,10 @@ object AttemptParitySuite:
     Pair(classOf[repositories.RepositoryApi], classOf[repositories.RepositoryApi.Attempt]),
     Pair(classOf[repositories.access.RepositoryAccessApi], classOf[repositories.access.RepositoryAccessApi.Attempt]),
     Pair(
+      classOf[repositories.access.RepositoryProtectionApi],
+      classOf[repositories.access.RepositoryProtectionApi.Attempt],
+    ),
+    Pair(
       classOf[repositories.actions.RepositoryActionApi],
       classOf[repositories.actions.RepositoryActionApi.Attempt],
     ),